+++ src/components/admin/transfers/CreateForm.tsx
@@ -13,18 +13,18 @@ import { ApiErrors } from 'service/apiErrors'
 import { AlertStore } from 'stores/AlertStore'
 import { CampaignResponse } from 'gql/campaigns'
 import { useCreateTransfer } from 'service/transfer'
-import VaultSelect from 'components/vaults/VaultSelect'
+import VaultSelect from 'components/admin/vaults/VaultSelect'
 import GenericForm from 'components/common/form/GenericForm'
 import SubmitButton from 'components/common/form/SubmitButton'
-import CurrencySelect from 'components/currency/CurrencySelect'
+import CurrencySelect from 'components/common/currency/CurrencySelect'
 import FormTextField from 'components/common/form/FormTextField'
 import { TransferData, TransferInput, TransferResponse } from 'gql/transfer'
 
 import SelectDate from './custom/SelectDate'
 import { TransferStatus } from './TransferTypes'
-import CampaignSelect from '../campaigns/CampaignSelect'
+import CampaignSelect from '../../campaigns/CampaignSelect'
 import { useVaultsList } from 'common/hooks/vaults'
-import PersonSelect from 'components/person/PersonSelect'
+import PersonSelect from 'components/admin/person/PersonSelect'
 import { fromMoney, toMoney } from 'common/util/money'
 
 type Props = {
