+++ src/components/admin/transfers/EditForm.tsx
@@ -15,17 +15,17 @@ import { AlertStore } from 'stores/AlertStore'
 import { CampaignResponse } from 'gql/campaigns'
 import { endpoints } from 'service/apiEndpoints'
 import { useEditTransfer } from 'service/transfer'
-import VaultSelect from 'components/vaults/VaultSelect'
+import VaultSelect from 'components/admin/vaults/VaultSelect'
 import GenericForm from 'components/common/form/GenericForm'
 import SubmitButton from 'components/common/form/SubmitButton'
-import CurrencySelect from 'components/currency/CurrencySelect'
+import CurrencySelect from 'components/common/currency/CurrencySelect'
 import FormTextField from 'components/common/form/FormTextField'
 import { TransferData, TransferInput, TransferResponse } from 'gql/transfer'
 
 import SelectDate from './custom/SelectDate'
 import { TransferStatus } from './TransferTypes'
 import TransferStatusSelect from './TransferStatusSelect'
-import CampaignSelect from '../campaigns/CampaignSelect'
+import CampaignSelect from '../../campaigns/CampaignSelect'
 import { fromMoney, toMoney } from 'common/util/money'
 
 const dateParser = (date: Date | undefined) => {
